[BUGFIX] Prevent endless processing of images in backend
To prevent double processing of files with identical configuration
a cleanup of the configuration was introduced in #102164.

It was however forgotten to further work with the cleaned up
configuration, which is introduced with this fix.

Also use integer values instead of strings in FileReferenceContainer,
as logically these values should always be integer.

Releases: main, 13.4
Resolves: #107520
Change-Id: Iabc3f244748a8345afc3bc3ce4b39454c1bbf1c3
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/90765
Tested-by: Oli Bartsch <bo@cedev.de>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Oli Bartsch <bo@cedev.de>