[TASK] Replace "new" button for scheduler tasks in FormEngine
Editing a `tx_scheduler_task` also allowed
creating a new task via the "New" button in
the FormEngine doc header. However, since
#107526, the task type can no longer be
changed in FormEngine. Therefore, the "New"
button was hidden as it led to a broken
editing view.

Thanks to #107471, a dedicated wizard is
available for creating new tasks. This wizard
can therefore now be used in FormEngine.

Resolves: #107539
Related: #107471
Related: #107526
Releases: main
Change-Id: I77f8887dba0dba902cd2af6ae70844fd448f00a3
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/90798
Reviewed-by: Garvin Hicking <garvin@hick.ing>
Tested-by: Oli Bartsch <bo@cedev.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Garvin Hicking <garvin@hick.ing>
Reviewed-by: Oli Bartsch <bo@cedev.de>
Tested-by: Benni Mack <benni@typo3.org>