Revert "[DOCS] Correct permalinks in EXT:form"
This reverts commit 1b102268ebd64a1c398bce8f29391bcf71fa99e3.

Reason for revert: The permalink docs.typo3.org access
was broken for some time, because links like

  https://docs.typo3.org/permalink/typo3/cms-scheduler:developer-guide

inadvertently got shared, while actually links needed to point to

  https://docs.typo3.org/permalink/typo3-cms-scheduler:developer-guide

via HTTP.

This was recently fixed in the docs.typo3.org permalink tool:

  https://github.com/TYPO3GmbH/site-intercept/pull/239

This commit was made under the assumption, the permalinks should be
converted to their original "canonical" format, too.

However, the RST syntax parsing actually uses the `typo3/cms-scheduler`
notation to be parsed. Using a `typo3-cms-scheduler` notation there is
NOT recognized by the render-guides parsers. Thus, the links using that
resolving are not parsed with their proper interlink repositories.

Due to this, the change should be reverted to not let the documentation
rendering (and nightlies) fail. A change to render-guides must first be
made to also allow both syntaxes, before this patch can be implemented
in TYPO3 documentation files (or it can be left as it is currently).

Releases: main, 13.4
Resolves: #107552
Reverts: #107519
Change-Id: Ia818d91ecab7bc98da529bbc2dc6d5e0f811ae0b
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/90818
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Garvin Hicking <gh@faktor-e.de>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Garvin Hicking <gh@faktor-e.de>