[TASK] Check ReST rendering for commits with .rst files
The nightlies already check the ReST rendering of the full
documentation files since #107514.

However, with a commit like #107519 it became obvious
that ReST rendering checks should better be performed
on pre-merge to avoid wrongly rendered documentation files,
which are directly published on docs.typo3.org after
merge.

The whole rendering process is a step that takes a
little while longer, so unconditionally including it
in pre-merge pipeline steps would take CI longer
than necessary.

This commit introduces pre-merge checking with a rendering
step, but only if a `*.rst` file is included in a commit.

This is done with a `rules` entry which does not allow
to be used in conjunction with `except`, which is why
the current checks are "reworded" to be put into the
`rules` entry.

Resolves: #107557
Related: #107514
Releases: main, 13.4, 12.4
Change-Id: Ic6cf80238d6627f1a32eac7a4623f1e841911c0e
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/90843
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: core-ci <typo3@b13.com>