[TASK] Use attribute selectors instead of pseudo-classes for disabled and readonly
Replace all occurrences of :disabled and :readonly with their
corresponding attribute selectors ([disabled] and [readonly]).

Pseudo-class selectors are handled internally by browsers and cannot
be implemented in custom web components. Since the state of these
elements is always reflected in their attributes, this change has no
impact on current implementations and serves as a preparation for
future work.

Resolves: #107585
Releases: main, 13.4
Change-Id: I4ef343f6ab5a50cdbcd20c0d9e5cc263296207b1
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/90880
Reviewed-by: Benjamin Kott <benjamin.kott@outlook.com>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benjamin Kott <benjamin.kott@outlook.com>