[BUGFIX] Prevent infinite loop in scheduler invalid task deletion
Deleting a task results in checking the related
task for invalidity. If invalid, the task should
be set to disabled. However, this DH operation
resulted in calling the scheduler related DH
hook which triggered the same operation again.

The handling is adjusted to set the disable
the task only in case not already disabled
and to use a simple update statement instead
of triggering DH.

Resolves: #107597
Releases: main
Change-Id: Ibc7d9c2d5d0e376072454f00e1bbb69131d9258e
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/90904
Reviewed-by: Garvin Hicking <garvin@hick.ing>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Lina Wolf <112@linawolf.de>
Reviewed-by: Josua Vogel <j.vogel97@web.de>
Tested-by: Oli Bartsch <bo@cedev.de>
Reviewed-by: Oli Bartsch <bo@cedev.de>
Tested-by: Lina Wolf <112@linawolf.de>
Tested-by: Garvin Hicking <garvin@hick.ing>