[TASK] Update `alwan` to 2.2.0
Update `alwan`, the used color picker library, to version 2.2.0.
This version contains markup and CSS adjustments and we strive to
use the upstream CSS from now on in order to avoid having to cope
with manual markup and CSS changes in future minor or patch updates.

Executed commands:

  npm install --save alwan@^2.2
  npm run build

Releases: main, 13.4
Resolves: #107616
Change-Id: Ib5981f3714412e23e4b0f5418179ecc843a5a7c9
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/91095
Reviewed-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benjamin Franzke <ben@bnf.dev>