[TASK] Enable EnvironmentStatusReport for CLI
Having the EnvironmentStatusReport disabled for CLI
makes the report scheduler task useless.

This patch allows to utilize the extended checks.

An important differentiation is that the environment
in CLI can be different from the Web configuration,
and certain variables are not available in CLI context
($_SERVER mostly).

This patch excludes these specific checks, and
should catch all places where this superglobal is accessed.

Additionally, to make recipients of the CLI-context based
report aware of possible differences between PHP configuration
on CLI and web context, a small disclaimer is added to
the email.

The new language service system has been utilized in the
places this touches.

Resolves: #94365
Releases: main, 13.4
Change-Id: I3d90b7822d0df86160a24afbf1f1aaf34414e7ac
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/93052
Tested-by: core-ci <typo3@b13.com>
Tested-by: Garvin Hicking <garvin@hick.ing>
Reviewed-by: Garvin Hicking <garvin@hick.ing>