[BUGFIX] Avoid applying CSP nonce sources when not required
This change aims to avoid applying Content-Security-Policy nonce
source when they are not actually required.

The `ConsumableNonce` class was refactored and does not extend
`ConsumableString` anymore. As a result the invocation counters
are now split into `consumeInline` (required usage of the nonce)
and `consumeStatic` (optional usage of the nonce) and are tracking
the different usage aspects, like their directives.

$nonce = new ConsumableNonce();
$nonce->consumeInline(Directive::ScriptSrcElem); // required usage
$nonce->consumeStatic(Directive::StyleSrcElem); // optional usage

Nonce sources are removed from the CSP policy in case
- the response content is readable and is empty (0 bytes); or
- the nonce consumption counter (all usages) is zero; or
- the consumed nonce sources for a particular directive can be
  avoided due to the existence of a source keyword exception
  (like using `'unsafe-inline'`); or
- the new `PolicyPreparedEvent` is used to programatically adjust
  the corresponding behavior to deny or allow  nonce sources

In case the nonce sources are supposed to be denied, both
`ContentSecurityPolicyHeaders` middlewares in the frontend
and backend scope are removing the corresponding nonce-related
literals from the generated HTML content of the response object.

Resolves: #107062
Releases: main, 13.4
Change-Id: I2de6d67e2639a5701c3a2acb0345e2c9c9b683de
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/90814
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Oliver Hader <oliver.hader@typo3.org>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Oliver Hader <oliver.hader@typo3.org>