[TASK] Avoid unintentional null values passed to TagBuilder
With Fluid 5, the behavior of `null` values passed to the `TagBuilder`
will change: Before, `null` was converted to string and thus lead to an
empty tag attribute. With Fluid 5, the attribute will be omitted
altogether.

While this is unproblematic in most circumstances, for the `alt`
attribute of image tags it isn't. Both `ImageViewHelper` and
`MediaViewHelper` are adjusted to always pass a string to that
attribute, which ensures that it will always be present on the resulting
`<img>`.

Resolves: #107660
Releases: main
Change-Id: I3007d7342bb9c3360eca7486cceaa54ff36c36c1
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/90998
Tested-by: Garvin Hicking <garvin@hick.ing>
Reviewed-by: Garvin Hicking <garvin@hick.ing>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Simon Praetorius <simon@praetorius.me>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Simon Praetorius <simon@praetorius.me>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>