[TASK] Remove experimental note in Fluid-based Page Module components
Several classes still contain a comment that they are
experimental and subject to change in TYPO3 10 or 11.

This change removes the experimental note because the
APIs can be considered stable. The `@internal` declaration
is kept because affected classes are not public API and
should not be extended or overwritten by developers.

Resolves: #107671
Releases: main
Signed-off-by: Torben Hansen <derhansen@gmail.com>
Change-Id: I80b02826f26961c24d30eb0a661b1e255275bf2b
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/91010
Tested-by: Josua Vogel <j.vogel97@web.de>
Reviewed-by: Josua Vogel <j.vogel97@web.de>
Tested-by: Oliver Klee <typo3-coding@oliverklee.de>
Tested-by: Lina Wolf <112@linawolf.de>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Reviewed-by: Lina Wolf <112@linawolf.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>