[TASK] Remove TSFE->cObj and TSFE->newCObj()
TypoScriptFrontentController->cObj is one of the last hidden loop
dependencies between TSFE, ContentObjectRenderer and Request: This
public property is initialized in a middleware with the request
and TSFE itself is then registered as global. This hides away a
dependency to Request. The patch dissolves the situation.

ext:form is (of course) a victim again and actively uses the
registered global to render content objects in a few cases where
Request is not necessarily available. The patch turns this around
and makes the handling more explicit.

The existing "boot TSFE" related hacks in ext:redirects receive
another ugly quirk to deal with this. Chances are this can be
simplified quite a bit when TSFE is finally gone, though.

In turn, TSFE->cObj and TSFE->newCObj() are finally gone, which
is another step towards better top->bottom logic, making TSFE
less godlike.

Resolves: #107698
Related: #96659
Releases: main
Change-Id: I02a281e745aecc10e5852c801792401ed259fa03
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/91064
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Benni Mack <benni@typo3.org>