[TASK] Avoid tsfe->recordRegister and tsfe->currentRecord
These two properties stem from ancient thinking: Their sole
usages are a combination of CONTENT and RECORDS content
objects where CONTENT may suppress rendering an element a
second time if it has already been processed on the page.

This has always been questionable:
First, 'insert record' (ctype shorctut) does not allow
selecting an element twice. And if an editor fiddles with
multiple 'insert record' elements on a page, it can expect
the selected records to be always rendered if they are
valid. supressing a second rendering leads to more harm
than good and is contraintuitive.
Additionally, content elements are nowadays more often
rendered using FLUIDTEMPLATE than using CONTENT, the
combination CONTENT<->RECORD usually doesn't kick in anymore.
And when CONTENT and RECORD are used as "native" TypoScript
in a page TS setup to render something, supressing
rendering of an element just because some other TS snipped
used that record before at a different place in the page
already is again more harmful than good. In this case for
the poor integrator that has to track down why some TS
snippet does not work as expected.

All in all, the patch removes the logic entirely.

Resolves: #107704
Releases: main
Change-Id: I66a6183ca67f2eb5c82057ae1b72e1d872f8eca5
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/91076
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Benni Mack <benni@typo3.org>