[BUGFIX] Preserve current folder in element browser
Creating new folders in the FileList
is using the element browser while
forwarding the currently active folder.

This folder is now correctly selected
in the file storage tree inside the
element browser. Since the event listener
are directly added to the component
instead of inside another listener.

Additionally, this now also works in
case the tree is initially collapsed
(slected node not loaded).

Resolves: #107705
Releases: main, 13.4
Change-Id: Icb908c4e370063da71f35a2b1eb3a23c72897c42
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/91077
Reviewed-by: Garvin Hicking <garvin@hick.ing>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Oli Bartsch <bo@cedev.de>
Tested-by: Oli Bartsch <bo@cedev.de>
Tested-by: Garvin Hicking <garvin@hick.ing>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: core-ci <typo3@b13.com>