[BUGFIX] Do not prefix URL in ImageResource
During the fix of #106002 it was overlooked, that ImageResource
object has two modes, which are checked later on in code.

One is when it holds a FAL object, where the URL is a full URL
including the correct prefix and the second is not holding such
object, where the prefix (absRefPrefix) is added at a later point
in code.

Therefore, when constructing the public URL from ImageInfo,
we must not prefix the resulting URL.

Resolves: #107758
Related: #106002
Releases: main, 13.4
Change-Id: Id56da2c2924ce0f3cd6f30f9a2992e3c2de78178
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/91152
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Benni Mack <benni@typo3.org>