[FEATURE] Support `username` for authentication in Redis cache backend
Starting with redis 6.0 authentication with `username & password` tuple
is supported, which was not reflected directly as configuration for the
`RedisBackend` cache backend of TYPO3.

Due to missing type on `RedisBackend->setPassword()` and the proeperty
`$password` instance maintainer configured the authentication tuple as
array for `password` as a workaround, which is quite etablished within
the TYPO3 community and hinders directly strict typing the places with
simple string type. Note that the annotated doc-block type `string` is
intentionally ignored for this working workaround.

This change provides now official support to configure `username` and
`password` for the redis connection of the cache backend while keeping
the workaround to configure a array as `password`, which is deprecated
now.

Resolves: #107725
Releases: main
Change-Id: I87da42018e39b5c6b9e23343d35bf75965950830
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/91151
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Josua Vogel <j.vogel97@web.de>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Sven Liebert <mail@sven-liebert.de>
Reviewed-by: Josua Vogel <j.vogel97@web.de>
Reviewed-by: Sven Liebert <mail@sven-liebert.de>
Tested-by: Guido Schmechel <guido.schmechel@brandung.de>
Reviewed-by: Guido Schmechel <guido.schmechel@brandung.de>