[DOCS] Avoid using the constant $GLOBALS['LANG']
At https://docs.typo3.org/permalink/t3coreapi:extension-localization-php we write:

The :php:`\TYPO3\CMS\Core\Localization\LanguageService` could formerly be
accessed via the global variable :php:`$GLOBALS['LANG']`. This is now
discouraged. Use :php:`LanguageServiceFactory` instead.

So we should not use the discouraged practice in new changelog entries. PPl are prone to copy & paste. And the constant is expecially misleading as it is available while a backend user is logged in so you might not notice right away that it is actually not working in frontend context.