[DOCS] Avoid using the constant $GLOBALS['LANG']
At https://docs.typo3.org/permalink/t3coreapi:extension-localization-php
we write:

The :php:`\TYPO3\CMS\Core\Localization\LanguageService` could formerly be
accessed via the global variable :php:`$GLOBALS['LANG']`. This is now
discouraged. Use :php:`LanguageServiceFactory` instead.

So we should not use the discouraged practice in new changelog entries.
PPl are prone to copy & paste. And the constant is expecially misleading
as it is available while a backend user is logged in so you might not
notice right away that it is actually not working in frontend context.

Releases: main
Resolves: #107781
Change-Id: Ifa0ae5e0734f2bb48d5e06111e8a35f623506871
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/91159
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>