[BUGFIX] FalStatus must only check browsable storages
The selection of storages to validate previously only
considered the "online"-status of a storage.
But, any storage may be online, but not browsable.

Since the FalStatus check is about the files
inside the storage, it does not make sense
(and actually fails) to check non-browsable storages.

Ensure only browsable storages are considered for this
status report.

Resolves: #107805
Releases: main, 13.4
Change-Id: Idd84b50d7769b8fbe1648754d7217d2eb8659349
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/91195
Tested-by: Johannes Kasberger <johannes.kasberger@reelworx.at>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Johannes Kasberger <johannes.kasberger@reelworx.at>
Tested-by: Oli Bartsch <bo@cedev.de>
Reviewed-by: Oli Bartsch <bo@cedev.de>