[TASK] Make GreedyDatabaseBackend more sophisticated
This patch improves GreedyDatabaseBackend, so that it
doesn't store the same database rows all over again.
This is done by remembering the uids of the result
rows and pointing them to the actual cache identifier.
The next time a uid from the last result is requested,
the pointer will be retrieved and with it the rows itself.

Resolves: #107819
Releases: main, 13.4
Change-Id: Iaa7ab850ed66ac2e24f77bc27c4eff293823b56e
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/91237
Reviewed-by: Simon Schaufelberger <simonschaufi+typo3@gmail.com>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Garvin Hicking <garvin@hick.ing>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Garvin Hicking <garvin@hick.ing>