[TASK] Declare ContentObjecRenderer stdWrapOrder constant
This property is not and should never be written. It is
now redeclared constant, protected and renamed.

The motivation for this is that ContentObjectRenderer
is stored as serialized string in cache_pages as part
of the rendering state for each non cached content
element ("INT" objects). Having this rather huge array
declared constant avoids this state being part of the
serialized string, making cache_pages rows smaller.

Resolves: #107835
Releases: main
Change-Id: I50a4ab59a0e7d81c15387f53a542414ec170dd39
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/91248
Reviewed-by: Garvin Hicking <garvin@hick.ing>
Tested-by: Oliver Klee <typo3-coding@oliverklee.de>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Garvin Hicking <garvin@hick.ing>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Benni Mack <benni@typo3.org>