[TASK] Add option to not add cache busting to generated URIs
The initial idea was to omit an option whether to add cache
busting for a generated URL or not and apply cache busting
everywhere. The reason for that was, that such specific option
might not be useful for other implementations of the URI generation
interface.

It turned out however, that there are legitimate cases where
not adding cache busting is required. Some of those are in code
that was touched already. For those a workaround for the missing
option was added, which was removing the query string. This workaround
is incomplete and also does not work when configuring versionNumberedFileNames.

Additionally, when generating URLs to JS assets, that already have a
hashed identifier in filename e.g. from a frontend build, adding
additional cache busting would not only be useless, it can also lead to
unnecessary additional requests, when the JS code itself required the
same asset without TYPO3 cache busting.

On top of that, especially when imagining resources publishing to CDNs
and generating URLs for those, not only the cache busting option would
be obsolete, but also the "absolute" option, as URLs would have to be
always absolute.

Also the prefix option would be obsolete.

Last but not least, by implementing this option, the impact of
some changes could be lowered as well, having less required deprecations
and less breaking changes.

Resolves: #107869
Releases: main
Change-Id: I44529a85fe4594c45de636989eb97ed921d987dc
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/91287
Tested-by: Simon Praetorius <simon@praetorius.me>
Tested-by: Garvin Hicking <garvin@hick.ing>
Reviewed-by: Garvin Hicking <garvin@hick.ing>
Tested-by: Helmut Hummel <typo3@helhum.io>
Reviewed-by: Helmut Hummel <typo3@helhum.io>
Reviewed-by: Simon Praetorius <simon@praetorius.me>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>