[BUGFIX] Detect `FIRST_INSTALL` in project path as well
In the past, a system administrator who set up TYPO3 had
to create a FIRST_INSTALL in the public path ("htdocs")
to gain access to the installer.

Since the introduction of the composer mode, it is possible
to have the public path in a subfolder as well.

However, people still needed to create the `FIRST_INSTALL`
file in the public path, whereas the project root path is
much more useful and convenient for newcomers.

This patch now:

* Checks for `FIRST_INSTALL` in both project root and
  public path, instead of only the public directory.

* Modified getFirstInstallFilePaths() to scan both paths
  and return the detected files with the absolute path.

* Updates `removeFirstInstallFile()` to delete found
  files using the now returned absolute path and not
  adding a prefix path anymore.

* Adds functional tests for new behavior.

Resolves: #108021
Releases: main, 13.4
Change-Id: I7426be6ae85824ef4c5da50c5de4e88ae50b9f15
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/91307
Reviewed-by: Garvin Hicking <garvin@hick.ing>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Garvin Hicking <garvin@hick.ing>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Stefan Bürk <stefan@buerk.tech>