[BUGFIX] Avoid performance bottleneck in system information toolbar
The system information toolbar renders the entire backend in the toolbar
dropdown, if logged out in another tab, because the request to the
information refresh route is redirected to the login page (and once
logged back in, to the main backend controller).

Once the main backend controller is inserted (e.g. after relogin in
another tab), a nested system information toolbar will be updated
to contain the backend again and again and again, resulting in
exponential growth, which can lead to 10GB of RAM usage and
4 million DOM nodes over time, making the entire backend unresponsive.

To circumvent this issue, we're now instruction `fetch()` to avoid
following redirects, which means we can detect that a redirect
would have performed and ignore the request result instead of
inserting invalid markup.

Releases: main, 13.4, 12.4
Resolves: #107894
Change-Id: I72ac8d3efc05c3c8e2d1e2dba0898515906b3e16
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/91311
Reviewed-by: Benjamin Kott <benjamin.kott@outlook.com>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benjamin Kott <benjamin.kott@outlook.com>
(cherry picked from commit 68f46b8cc541ab3fc1be063eddd17d2642ac4495)
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/91429
Tested-by: Garvin Hicking <garvin@hick.ing>
Reviewed-by: Garvin Hicking <garvin@hick.ing>