[TASK] Adapt translation domain in EXT:filemetadata
This patch adjusts translation domains within EXT:filemetadata
to use the new syntax.

In addition, the file is moved from locallang_tca.xlf
to db.xlf with a fallback layer in place for this
file if third-party extensions have referenced it, the
reference is continuing to work.

Related: #93334
Resolves: #107912
Releases: main
Change-Id: Ia9d167710e60d59043be8f197cac39743022a897
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/91341
Tested-by: Lina Wolf <112@linawolf.de>
Reviewed-by: Lina Wolf <112@linawolf.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>