[FEATURE] Show editor information in workspace overview
Add a new "Editor" column to the workspace module that displays
the username of the last editor who modified each record. The
editor information is retrieved from the record history and
displayed after the "Last changed on" column.

The column shows the editor's username in a badge, or "Unknown"
if no editor information is available. This helps reviewers
quickly identify who last worked on each workspace record.

Resolves: #106074
Releases: main
Change-Id: I62d95fef4942e59ace63e433944288215a958987
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/91343
Tested-by: Fabien Udriot <fabien@omic.ch>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Fabien Udriot <fabien@omic.ch>
Tested-by: Oli Bartsch <bo@cedev.de>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Oli Bartsch <bo@cedev.de>