[TASK] Adapt translation domain in EXT:impexp
This patch adjusts translation domains within EXT:impexp
to use the new syntax.

In addition, the file is moved from locallang_tca.xlf
to db.xlf with a fallback layer in place for this
file if third-party extensions have referenced it, the
reference is continuing to work.

Related: #93334
Resolves: #107916
Releases: main
Change-Id: I6dd7d5de0f445569575e04e724dacc8f355ec511
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/91344
Reviewed-by: Oli Bartsch <bo@cedev.de>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Oli Bartsch <bo@cedev.de>