[!!!][TASK] Remove FE & BE HTTP response compression
When `$GLOBALS['TYPO3_CONF_VARS']['BE']['compressionLevel']` or
`$GLOBALS['TYPO3_CONF_VARS']['FE']['compressionLevel']` are set
to a value between 1 and 9 (default 0), they enable gzip
compression of the HTTP application resonse in a BE or FE
middleware.

The patch removes this feature: It stems back to a time
where web servers didn't usally do that on their own and
applications had to take care. It is nowadays largely
unused in application layers since web servers are the
better place to apply HTTP compression and can use
improved compression standards like brotli or zStandard.

The FE implementation used a custom class to handle the
compression which is removed with the patch. In the backend,
the OutputCompression middleware was the sole use of toggle
`$GLOBALS['TYPO3_CONF_VARS']['BE']['compressionLevel']`, and
thus removed as well.

Resolves: #107943
Releases: main
Change-Id: I9e59f0d54f51c3f6c3092064b185b8608a396605
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/91378
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Markus Klein <markus.klein@typo3.org>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>