[BUGFIX] Define height for modal size "small"
Currently, when opening a modal of size "small"
using the iframe type, the modal height is 0px

This has been fixed by defining css height (same
as width) when a iframe is used. A max-height
is now also defined to ensure the complete modal
is visible in the viewport.

Resolves: #107962
Releases: main, 13.4
Change-Id: Icb2a10f90726e76017450f7ccb1071aa81f346ad
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/91403
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oli Bartsch <bo@cedev.de>
Reviewed-by: Oli Bartsch <bo@cedev.de>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: Benjamin Franzke <ben@bnf.dev>