[BUGFIX] Preserve restrictions for subselect in GreedyDatabaseBackend
When fetching rows in the GreedyDatabaseBackend,
the query can be customized via the VisibilityAspect.
Based on the aspect, query restrictions are removed.
Those modified restrictions must be preserved for
the subselect.

One place, which is fixed with this patch is the
content preview of "shortcut" records, which does
now display hidden relations again.

Resolves: #108034
Releases: main, 13.4
Change-Id: I7edd6f5d5ae4249c1f52bffd7438126a6e7b0e8b
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/91488
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Oli Bartsch <bo@cedev.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oli Bartsch <bo@cedev.de>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Jochen Roth <rothjochen@gmail.com>
Tested-by: Jochen Roth <rothjochen@gmail.com>